requirements.txt
numpy==1.25.1
Pillow==11.1.0
